Split out nav into separate component
